refactor: 잘못된 package 이름 수정 및 파일 이동
