+++ src/main/java/gift/exception/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package gift.product.exception;
+package gift.exception;
 
 public class ProductNotFoundException extends RuntimeException {
     public ProductNotFoundException(Long id) {
