+++ src/main/java/gift/product/repository/ProductRepository.java
@@ -1,7 +1,7 @@
 package gift.product.repository;
 
 import gift.product.entity.Product;
-import gift.product.exception.ProductNotFoundException;
+import gift.exception.ProductNotFoundException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.JdbcClient;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
