+++ src/main/java/gift/product/service/ProductService.java
@@ -3,7 +3,7 @@ package gift.product.service;
 import gift.product.dto.request.ProductRequestDto;
 import gift.product.dto.response.ProductResponseDto;
 import gift.product.entity.Product;
-import gift.product.exception.ProductNotFoundException;
+import gift.exception.ProductNotFoundException;
 import gift.product.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import java.util.List;
